feat: Millard's model notebook
Included the Jupyter notebook to apply the PINN on Millard's model. The notebook creates a CSV file with generated data from the ODE system. Modified the random_ranges function in tools to allow for wider ranges. Enabled the user to impact the residual loss by providing a coef_res list to initialize the PINN. This ensures that the residual loss is always included in the total loss (no more comparisons required).