feat: multi-loss balancing
Cleaned the code and added the possibility of using different methods to balance the loss (packed in a function in tools). Changed the progress bars during training. The optimizer and scheduler are now defined inside of the pinn rather than in the notebooks. New way of saving results to allow taking into account more hyperparameters.